+++ Assets/Scripts/AI/BehaviourTree/BehaviourTree.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using AI.BehaviourTree.CoreNodes;
 
 #if UNITY_EDITOR
 using UnityEditor;
