+++ Assets/Scripts/AI/BehaviourTree/CoreNodes/RepeatNode.cs
@@ -1,4 +1,4 @@
-namespace AI.BehaviourTree
+namespace AI.BehaviourTree.CoreNodes
 {
     public class RepeatNode : DecoratorNode
     {
