+++ Assets/Scripts/AI/BehaviourTree/CoreNodes/RootNode.cs
@@ -1,0 +1,23 @@
+namespace AI.BehaviourTree.CoreNodes
+{
+    public class RootNode : Node
+    {
+        public Node child;
+    
+        protected override void OnStart()
+        {
+        }
+
+        protected override void OnStop()
+        {
+        }
+
+        protected override State OnUpdate() => child.Update();
+        public override Node Clone()
+        {
+            RootNode node = Instantiate(this);
+            node.child = child.Clone();
+            return node;
+        }
+    }
+}
