+++ Assets/Scripts/AI/BehaviourTree/CoreNodes/SequencerNode.cs
@@ -1,4 +1,4 @@
-namespace AI.BehaviourTree
+namespace AI.BehaviourTree.CoreNodes
 {
     public class SequencerNode : CompositeNode
     {
