+++ Assets/Scripts/AI/BehaviourTree/CoreNodes/WaitNode.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 541358dbdf081964c8ee478c23d64e03
+guid: 8a46a9fe77c3703489a8ba29441da955
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
