--- Assets/Scripts/AI/BehaviourTree/RepeatNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8db426307ec880446976437da16091d6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
