--- Assets/Scripts/AI/BehaviourTree/WaitNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 312c7a4b11b0b07408dc041eeaf99473
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
