+++ Assets/Scripts/Editor/BehaviourTreeView.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using AI.BehaviourTree;
+using AI.BehaviourTree.CoreNodes;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
 using UnityEngine;
