+++ Assets/Scripts/Editor/NodeView.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using AI.BehaviourTree;
+using AI.BehaviourTree.CoreNodes;
 using Unity.VisualScripting;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
